Now playing update

+++ src/events/trackEnd.js
@@ -12,14 +12,16 @@ class TrackEnd extends Event {
     let player = this.client.players.get(guild.id)
     if (!player) return
 
-    await player.queue.nowPlayingMessage
-      .delete()
-      .catch((err) => this.logger.error(err))
-
-    player.queue.nowPlayingMessage = null
-
     if (player.queue.tracks.length == 0) {
       this.client.emit('trackQueueEnd', player.queue.interactionChannel)
+
+      if (player.queue.nowPlayingMessage) {
+        await player.queue.nowPlayingMessage
+          .delete()
+          .catch((err) => this.logger.error(err))
+  
+        player.queue.nowPlayingMessage = null
+      }
     }
 
     this.client.user.setPresence({
